backend_flask_app/requirements.txt
aniso8601==7.0.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
promise==2.3
protobuf==5.28.1
PyJWT==2.9.0
PyMatting==1.1.12
pymongo==4.9.1
pyreadline3==3.5.3
pytest==8.3.3
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
Rx==1.6.3
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sympy==1.13.2
tifffile==2024.8.30
tqdm==4.66.5
urllib3==2.2.3
Werkzeug==3.0.4
